+++ src/Simpletools/Db/Cassandra/Result.php
@@ -320,7 +320,7 @@ class Result implements \Iterator
         return isset($this->_firstRowCache->{$name});
     }
 
-    public function rewind()
+    public function rewind() : void
     {
         $this->_result->rewind();
         $this->_position 	= 0;
@@ -331,24 +331,24 @@ class Result implements \Iterator
         }
     }
 
-    public function current()
+    public function current() : mixed
     {
         return $this->_currentRow;
     }
 
-    public function key()
+    public function key() : mixed
     {
         return $this->_position;
     }
 
-    public function next()
+    public function next() : void
     {
         $this->_currentRow = $this->fetch();
         ++$this->position;
-        return $this->_currentRow;
+        // return $this->_currentRow;
     }
 
-    public function valid()
+    public function valid() : bool
     {
         return ($this->_currentRow) ? true : false;
     }