+++ src/Simpletools/Db/Cassandra/Type/BigInt.php
@@ -21,7 +21,7 @@ class BigInt implements \JsonSerializable
         return $this->_value;
     }
 
-		public function jsonSerialize()
+		public function jsonSerialize() : mixed
 		{
 			return (string)$this->value();
 		}