+++ src/Simpletools/Db/Cassandra/Type/Blob.php
@@ -23,7 +23,7 @@ class Blob implements \JsonSerializable
 		return $this->_value;
 	}
 
-	public function jsonSerialize()
+	public function jsonSerialize() : mixed
 	{
 		return (int) $this->value();
 	}