+++ src/Simpletools/Db/Cassandra/Type/Inet.php
@@ -23,7 +23,7 @@ class Inet implements \JsonSerializable
 		return $this->_value;
 	}
 
-	public function jsonSerialize()
+	public function jsonSerialize() : mixed
 	{
 		return (int)$this->value();
 	}