+++ src/Simpletools/Db/Cassandra/Type/Set.php
@@ -42,7 +42,7 @@ class Set implements \JsonSerializable
 		return $this->_value;
 	}
 
-	public function jsonSerialize()
+	public function jsonSerialize() : mixed
 	{
 		return $this->_body;
 	}