+++ src/Simpletools/Db/Cassandra/Type/SimpleFloat.php
@@ -19,7 +19,7 @@ class SimpleFloat implements \JsonSerializable
 		return $this->_value;
 	}
 
-	public function jsonSerialize()
+	public function jsonSerialize() : mixed
 	{
 		return (float) $this->value();
 	}