+++ src/Simpletools/Db/Cassandra/Type/Tinyint.php
@@ -20,7 +20,7 @@ class Tinyint implements \JsonSerializable
 		return $this->_value;
 	}
 
-	public function jsonSerialize()
+	public function jsonSerialize() : mixed
 	{
 		return (int) $this->value();
 	}