+++ src/Simpletools/Db/Cassandra/Type/Uuid.php
@@ -24,7 +24,7 @@ class Uuid implements \JsonSerializable
         return $this->_value;
     }
 
-    public function jsonSerialize()
+    public function jsonSerialize() : mixed
     {
         return (string)$this->value();
     }